refactor(translate): move CanonicalResponsesPayload into @floway-dev/translate and return it directly from *-via-responses translators
`*-via-responses` translators always construct their target's `input` as
`ResponsesInputItem[]`; the wire type's `string | ResponsesInputItem[]`
union captured a branch these translators cannot produce. The gateway
had to lift each translated payload back into its narrower canonical
form via a `canonicalizeResponsesPayload(translated)` wrap at every
attempt call site.

Move `CanonicalResponsesPayload` and the wire-to-canonical
`canonicalizeResponsesPayload` helper into
`@floway-dev/translate/via-responses/responses-items` alongside the
existing `responsesItemsView`. Each `translate*ViaResponses` trip
declares its target as `CanonicalResponsesPayload`; the three gateway
attempt sites drop the wrap since the translator's return type IS
canonical.

`canonicalizeResponsesPayload` still runs at the two wire boundaries
that produce a raw `ResponsesPayload`: the Responses HTTP entry parsing
a request body, and the Responses WebSocket entry building a payload
from a client frame. Both now import the helper from translate.

Gateway consumers of `CanonicalResponsesPayload` (interceptors, tests,
rewrite, serve/serve-prep/attempt, image-generation server tool) all
switch their import from the local `interceptors/types.ts` to
`@floway-dev/translate/via-responses/responses-items`, and
`interceptors/types.ts` collapses to the app-side
`ResponsesInterceptor`, `ResponsesInvocation`, and
`ResponsesAttemptResult` types — with the canonical payload imported
from translate.
