+++ packages/gateway/src/data-plane/chat/responses/interceptors/compact-shim_test.ts
@@ -1,7 +1,7 @@
 import { test } from 'vitest';
 
 import { expandShimCompactionItems, withResponsesCompactShim } from './compact-shim.ts';
-import type { CanonicalResponsesPayload, ResponsesInvocation } from './types.ts';
+import type { ResponsesInvocation } from './types.ts';
 import { encodeBase64UrlJson } from '../../../../shared/base64url-json.ts';
 import type { ChatGatewayCtx } from '../../shared/gateway-ctx.ts';
 import { createNonResponsesSourceStore } from '../items/store.ts';
@@ -9,6 +9,7 @@ import { doneFrame, eventFrame, type ProtocolFrame } from '@floway-dev/protocols
 import { collectResponsesProtocolEventsToResult, type ResponsesPayload, type ResponsesResult, type ResponsesStreamEvent } from '@floway-dev/protocols/responses';
 import { eventResult, type ExecuteResult } from '@floway-dev/provider';
 import { assertEquals, stubModelCandidate, testTelemetryModelIdentity } from '@floway-dev/test-utils';
+import type { CanonicalResponsesPayload } from '@floway-dev/translate/via-responses/responses-items';
 
 const stubCtx: ChatGatewayCtx = {
   apiKeyId: 'test-key',
