+++ packages/gateway/src/data-plane/chat/responses/interceptors/types.ts
@@ -2,32 +2,9 @@ import type { TokenUsage } from '../../../../repo/types.ts';
 import type { ChatGatewayCtx } from '../../shared/gateway-ctx.ts';
 import type { Interceptor } from '@floway-dev/interceptor';
 import type { ProtocolFrame } from '@floway-dev/protocols/common';
-import type { ResponsesInputItem, ResponsesPayload, ResponsesResult, ResponsesStreamEvent } from '@floway-dev/protocols/responses';
+import type { ResponsesResult, ResponsesStreamEvent } from '@floway-dev/protocols/responses';
 import type { ExecuteResult, ResponsesInvocation as WireResponsesInvocation, TelemetryModelIdentity } from '@floway-dev/provider';
-
-// Wire `ResponsesPayload.input` is `string | InputItem[]`. The HTTP/WS entries
-// normalize the wire string to a single synthetic user-role message before
-// any serve logic runs, so every reader past the entry sees array shape.
-// This is the gateway-internal narrowed view; the wire-shape `ResponsesPayload`
-// continues to live on the entry boundary, on provider-call bodies, and in
-// translate-package signatures (those interface with foreign wire shapes that
-// haven't been normalized).
-export type CanonicalResponsesPayload = Omit<ResponsesPayload, 'input'> & {
-  input: ResponsesInputItem[];
-};
-
-// Lifts a wire `ResponsesPayload` into the gateway's internal canonical form
-// by replacing a bare-string `input` with a single synthetic user-role
-// message. Used at every boundary that produces a wire payload destined for
-// the gateway's serve/attempt pipeline: the HTTP/WS entries on a fresh
-// request, and the translate package's `messages → responses` /
-// `gemini → responses` outputs flowing into `responsesAttempt`.
-export const canonicalizeResponsesPayload = (payload: ResponsesPayload): CanonicalResponsesPayload => ({
-  ...payload,
-  input: typeof payload.input === 'string'
-    ? [{ type: 'message', role: 'user', content: payload.input }]
-    : payload.input,
-});
+import type { CanonicalResponsesPayload } from '@floway-dev/translate/via-responses/responses-items';
 
 export interface ResponsesInvocation extends Omit<WireResponsesInvocation, 'payload'> {
   payload: CanonicalResponsesPayload;
