+++ packages/gateway/src/data-plane/chat/responses/interceptors/vendor-deepseek-normalize.ts
@@ -18,7 +18,8 @@
 // Reference:
 // - https://api-docs.deepseek.com/zh-cn/guides/thinking_mode
 
-import type { CanonicalResponsesPayload, ResponsesInterceptor } from './types.ts';
+import type { ResponsesInterceptor } from './types.ts';
+import type { CanonicalResponsesPayload } from '@floway-dev/translate/via-responses/responses-items';
 
 interface DeepseekDisableField {
   thinking?: { type: 'disabled' };
