+++ packages/gateway/src/data-plane/chat/responses/serve.ts
@@ -1,10 +1,11 @@
 import { responsesAttempt } from './attempt.ts';
-import type { CanonicalResponsesPayload, ResponsesAttemptResult } from './interceptors/types.ts';
+import type { ResponsesAttemptResult } from './interceptors/types.ts';
 import { prepareResponsesServePlan } from './serve-prep.ts';
 import type { ChatGatewayCtx } from '../shared/gateway-ctx.ts';
 import type { ProtocolFrame } from '@floway-dev/protocols/common';
 import type { ResponsesStreamEvent } from '@floway-dev/protocols/responses';
 import type { ExecuteResult } from '@floway-dev/provider';
+import type { CanonicalResponsesPayload } from '@floway-dev/translate/via-responses/responses-items';
 
 export interface ResponsesServeGenerateArgs {
   readonly payload: CanonicalResponsesPayload;
